+++ apps/service/src/features/profile-container/api/index.ts
@@ -1,3 +1,4 @@
 export { default as fetchUserProfile } from './fetchUserProfile'
 export { default as postDuplicateName } from './postDuplicateName'
+export { default as postImgUrl } from './postImgUrl'
 export { default as postUserSettingForm } from './postUserSettingForm'
